crud_test.go
package main

import (
	"net/http"
	"net/http/httptest"
	"strings"
	"testing"

	"github.com/gin-gonic/gin"
	"github.com/stretchr/testify/assert"
)

func TestCreateComponentHandler(t *testing.T) {
	// Create a new Gin router and set the handler
	r := gin.Default()
	componentService := &ComponentService{} // Assuming you have initialized ComponentService correctly
	r.POST("/create", componentService.CreateComponent)

	// Test case: Successful creation
	payload := `{"name": "Component A", "description": "This is component A"}`
	req, _ := http.NewRequest("POST", "/create", strings.NewReader(payload))
	req.Header.Set("Content-Type", "application/json")
	resp := httptest.NewRecorder()
	r.ServeHTTP(resp, req)

	assert.Equal(t, http.StatusCreated, resp.Code)
	// Additional assertions based on the response body or any other expected behavior

	// Test case: Invalid JSON payload
	req, _ = http.NewRequest("POST", "/create", strings.NewReader("invalid-json"))
	req.Header.Set("Content-Type", "application/json")
	resp = httptest.NewRecorder()
	r.ServeHTTP(resp, req)

	assert.Equal(t, http.StatusBadRequest, resp.Code)
	// Additional assertions based on the response body or any other expected behavior
}

func TestReadComponentHandler(t *testing.T) {
	// Create a new Gin router and set the handler
	r := gin.Default()
	componentService := &ComponentService{} // Assuming you have initialized ComponentService correctly
	r.GET("/read/:id", componentService.ReadComponent)

	// Test case: Component found
	req, _ := http.NewRequest("GET", "/read/1", nil)
	resp := httptest.NewRecorder()
	r.ServeHTTP(resp, req)

	assert.Equal(t, http.StatusOK, resp.Code)
	// Additional assertions based on the response body or any other expected behavior

	// Test case: Component not found
	req, _ = http.NewRequest("GET", "/read/non_existent_id", nil)
	resp = httptest.NewRecorder()
	r.ServeHTTP(resp, req)

	assert.Equal(t, http.StatusNotFound, resp.Code)
	// Additional assertions based on the response body or any other expected behavior
}

func TestReplaceComponentHandler(t *testing.T) {
	// Create a new Gin router and set the handler
	r := gin.Default()
	componentService := &ComponentService{} // Assuming you have initialized ComponentService correctly
	r.PUT("/replace", componentService.ReplaceComponent)

	// Test case: Successful update
	payload := `{"id": 1, "name": "Updated Component A", "description": "This is the updated component A"}`
	req, _ := http.NewRequest("PUT", "/replace", strings.NewReader(payload))
	req.Header.Set("Content-Type", "application/json")
	resp := httptest.NewRecorder()
	r.ServeHTTP(resp, req)

	assert.Equal(t, http.StatusOK, resp.Code)
	// Additional assertions based on the response body or any other expected behavior

	// Test case: Invalid JSON payload
	req, _ = http.NewRequest("PUT", "/replace", strings.NewReader("invalid-json"))
	req.Header.Set("Content-Type", "application/json")
	resp = httptest.NewRecorder()
	r.ServeHTTP(resp, req)

	assert.Equal(t, http.StatusNotAcceptable, resp.Code)
	// Additional assertions based on the response body or any other expected behavior
}

func TestDeleteComponentHandler(t *testing.T) {
	// Create a new Gin router and set the handler
	r := gin.Default()
	componentService := &ComponentService{} // Assuming you have initialized ComponentService correctly
	r.DELETE("/delete/:id", componentService.DeleteComponent)

	// Test case: Successful deletion
	req, _ := http.NewRequest("DELETE", "/delete/1", nil)
	resp := httptest.NewRecorder()
	r.ServeHTTP(resp, req)

	assert.Equal(t, http.StatusOK, resp.Code)
	// Additional assertions based on the response body or any other expected behavior

	// Test case: Component not found
	req, _ = http.NewRequest("DELETE", "/delete/non_existent_id", nil)
	resp = httptest.NewRecorder()
	r.ServeHTTP(resp, req)

	assert.Equal(t, http.StatusNotFound, resp.Code)
	// Additional assertions based on the response body or any other expected behavior
}
